chore: Add user id in visual debug util (#14768)
